% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahut.R
\docType{package}
\name{datahut}
\alias{datahut}
\title{datahut: A data package for statistical analysis}
\description{
Provides datasets for Statistical Analysis
}
\author{
Priyanga Dilini Talagala, Rob J. Hyndman
}
