% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourist_sl.R
\docType{data}
\name{tourist_sl}
\alias{tourist_sl}
\title{Tourist arrivals to Sri Lanka}
\format{
Time series of class tsibble
\describe{
\item{Year}{Year}
\item{Arrivals}{Tourist arrivals}
}
}
\source{
Sri Lanka Annual Statistical Reports
}
\usage{
tourist_sl
}
\description{
Tourist arrivals to Sri Lanka from 1970 to 2018
}
\examples{
library(fable)
autoplot(tourist_sl)
}
\author{
Priyanga Dilini Talagala, Claire Kermorvant
}
\keyword{datasets}
