% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touristsl.R
\docType{data}
\name{touristsl}
\alias{touristsl}
\title{Tourist arrivals to Sri Lanka}
\format{
Time series of class tsibble
\describe{
\item{Year}{Year}
\item{Arrivals}{Tourist arrivals}
}
}
\source{
Annual Statistical Reports, Sri Lanka Tourism Development Authority
}
\usage{
touristsl
}
\description{
Tourist arrivals to Sri Lanka from 1970 to 2019.
There was a civil war in the country from July 1983 to May 2009.
Sri Lanka Easter bombings happened on 21 April 2019.
On 26 December 2004, Sri Lanka became a victim of the Indian Ocean Tsunami.
}
\examples{
library(fable)
autoplot(touristsl)
}
\author{
Priyanga Dilini Talagala
}
\keyword{datasets}
