% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touristsl_monthly.R
\docType{data}
\name{touristsl_monthly}
\alias{touristsl_monthly}
\title{Monthly tourist arrivals to Sri Lanka}
\format{
Time series of class tsibble
\describe{
\item{Month}{Year and Month}
\item{Arrivals}{Tourist arrivals}
}
}
\source{
Annual Statistical Reports, Sri Lanka Tourism Development Authority
}
\usage{
touristsl_monthly
}
\description{
Monthly tourist arrivals to Sri Lanka
}
\examples{
library(fable)
head(touristsl_monthly)
touristsl_monthly[433:588,]  \%>\% autoplot()
}
\author{
Priyanga Dilini Talagala, Claire Kermorvant
}
\keyword{datasets}
