% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touristsl.R
\docType{data}
\name{touristsl}
\alias{touristsl}
\title{Tourist arrivals to Sri Lanka}
\format{
Time series of class tsibble
\describe{
\item{Year}{Year}
\item{Arrivals}{Tourist arrivals}
}
}
\source{
Annual Statistical Reports, Sri Lanka Tourism Development Authority
}
\usage{
touristsl
}
\description{
Tourist arrivals to Sri Lanka from 1970 to 2019
}
\examples{
library(fable)
autoplot(touristsl)
}
\author{
Priyanga Dilini Talagala, Claire Kermorvant
}
\keyword{datasets}
