% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touristsl_monthly.R
\docType{data}
\name{touristsl_monthly}
\alias{touristsl_monthly}
\title{Monthly tourist arrivals to Sri Lanka}
\format{
Time series of class tsibble
\describe{
\item{Month}{Year and Month}
\item{Arrivals}{Tourist arrivals}
}
}
\source{
Annual Statistical Reports, Sri Lanka Tourism Development Authority
}
\usage{
touristsl_monthly
}
\description{
Monthly tourist arrivals to Sri Lanka.
There was a civil war in the country from July 1983 to May 2009.
Sri Lanka Easter bombings happened on 21 April 2019.
On 26 December 2004, Sri Lanka became a victim of the Indian Ocean Tsunami.
}
\examples{
library(fable)
library(magrittr)
head(touristsl_monthly)
touristsl_monthly[433:588,]  \%>\% autoplot()
}
\author{
Priyanga Dilini Talagala
}
\keyword{datasets}
